+++ packages/roosterjs-content-model-api/test/modelApi/table/deleteTableColumnTest.ts
@@ -1,6 +1,5 @@
-import { createTable, createTableCell } from 'roosterjs-content-model-dom';
+import { createTable, createTableCell, hasSelectionInBlock } from 'roosterjs-content-model-dom';
 import { deleteTableColumn } from '../../../lib/modelApi/table/deleteTableColumn';
-import { hasSelectionInBlock } from 'roosterjs-content-model-core';
 
 describe('deleteTableColumn', () => {
     it('empty table', () => {
