+++ packages/roosterjs-content-model-api/test/publicApi/block/setAlignmentTest.ts
@@ -1,4 +1,4 @@
-import * as normalizeTable from 'roosterjs-content-model-core/lib/publicApi/table/normalizeTable';
+import * as normalizeTable from 'roosterjs-content-model-dom/lib/modelApi/editing/normalizeTable';
 import { createContentModelDocument } from 'roosterjs-content-model-dom';
 import { IEditor } from 'roosterjs-content-model-types';
 import { paragraphTestCommon } from './paragraphTestCommon';
