+++ packages/roosterjs-content-model-api/test/publicApi/entity/insertEntityTest.ts
@@ -1,7 +1,7 @@
 import * as entityUtils from 'roosterjs-content-model-dom/lib/domUtils/entityUtils';
 import * as insertEntityModel from '../../../lib/modelApi/entity/insertEntityModel';
 import * as normalizeContentModel from 'roosterjs-content-model-dom/lib/modelApi/common/normalizeContentModel';
-import { ChangeSource } from 'roosterjs-content-model-core';
+import { ChangeSource } from 'roosterjs-content-model-dom';
 import { IEditor } from 'roosterjs-content-model-types';
 import { insertEntity } from '../../../lib/publicApi/entity/insertEntity';
 import {
