+++ packages/roosterjs-content-model-api/test/publicApi/image/changeImageTest.ts
@@ -1,4 +1,4 @@
-import * as readFile from 'roosterjs-content-model-core/lib/publicApi/domUtils/readFile';
+import * as readFile from 'roosterjs-content-model-dom/lib/domUtils/readFile';
 import { changeImage } from '../../../lib/publicApi/image/changeImage';
 import { IEditor } from 'roosterjs-content-model-types';
 import {
