+++ packages/roosterjs-content-model-core/test/coreApi/formatContentModel/formatContentModelTest.ts
@@ -1,6 +1,5 @@
-import * as transformColor from '../../../lib/publicApi/color/transformColor';
-import { ChangeSource } from '../../../lib/constants/ChangeSource';
-import { createImage } from 'roosterjs-content-model-dom';
+import * as transformColor from 'roosterjs-content-model-dom/lib/domUtils/style/transformColor';
+import { ChangeSource, createImage } from 'roosterjs-content-model-dom';
 import { formatContentModel } from '../../../lib/coreApi/formatContentModel/formatContentModel';
 import {
     ContentModelDocument,
