+++ packages/roosterjs-content-model-core/test/editor/core/createEditorDefaultSettingsTest.ts
@@ -4,7 +4,7 @@ import { tablePreProcessor } from '../../../lib/override/tablePreProcessor';
 import {
     listItemMetadataApplier,
     listLevelMetadataApplier,
-} from '../../../lib/metadata/updateListMetadata';
+} from '../../../lib/override/listMetadataApplier';
 import {
     createDomToModelSettings,
     createModelToDomSettings,
