+++ packages/roosterjs-content-model-core/test/overrides/pasteGeneralProcessorTest.ts
@@ -1,4 +1,4 @@
-import * as sanitizeElement from '../../lib/utils/sanitizeElement';
+import * as sanitizeElement from '../../lib/command/createModelFromHtml/sanitizeElement';
 import { createContentModelDocument } from 'roosterjs-content-model-dom';
 import { DomToModelContext } from 'roosterjs-content-model-types';
 import {
