+++ packages/roosterjs-content-model-dom/test/domToModel/processors/listProcessorTest.ts
@@ -1,11 +1,11 @@
 import * as stackFormat from '../../../lib/domToModel/utils/stackFormat';
-import { BulletListType } from 'roosterjs-content-model-core/lib/constants/BulletListType';
+import { BulletListType } from '../../../lib/constants/BulletListType';
 import { childProcessor as originalChildProcessor } from '../../../lib/domToModel/processors/childProcessor';
 import { createContentModelDocument } from '../../../lib/modelApi/creators/createContentModelDocument';
 import { createDomToModelContext } from '../../../lib/domToModel/context/createDomToModelContext';
 import { DomToModelContext, ElementProcessor } from 'roosterjs-content-model-types';
 import { listProcessor } from '../../../lib/domToModel/processors/listProcessor';
-import { NumberingListType } from 'roosterjs-content-model-core/lib/constants/NumberingListType';
+import { NumberingListType } from '../../../lib/constants/NumberingListType';
 
 describe('listProcessor', () => {
     let context: DomToModelContext;
