+++ packages/roosterjs-content-model-plugins/test/paste/processPastedContentFromWacTest.ts
@@ -7,7 +7,7 @@ import { processPastedContentWacComponents } from '../../lib/paste/WacComponents
 import {
     listItemMetadataApplier,
     listLevelMetadataApplier,
-} from 'roosterjs-content-model-core/lib/metadata/updateListMetadata';
+} from 'roosterjs-content-model-core/lib/override/listMetadataApplier';
 
 import {
     contentModelToDom,
