+++ packages/roosterjs-content-model-plugins/test/tableEdit/TableEditTestHelper.ts
@@ -1,7 +1,6 @@
 import * as TestHelper from '../TestHelper';
 import { DOMEventHandlerFunction } from 'roosterjs-editor-types';
-import { getObjectKeys } from 'roosterjs-content-model-dom';
-import { normalizeTable } from 'roosterjs-content-model-core';
+import { getObjectKeys, normalizeTable } from 'roosterjs-content-model-dom';
 import { TableEditPlugin } from '../../lib/tableEdit/TableEditPlugin';
 import {
     ContentModelTable,
