+++ packages/roosterjs-editor-adapter/lib/editor/utils/selectionConverter.ts
@@ -1,7 +1,7 @@
 import { createRange } from 'roosterjs-editor-dom';
-import { parseTableCells } from 'roosterjs-content-model-dom';
+import { createTableRanges } from './createTableRanges';
 import { SelectionRangeTypes } from 'roosterjs-editor-types';
-import type { DOMSelection, TableSelection } from 'roosterjs-content-model-types';
+import type { DOMSelection } from 'roosterjs-content-model-types';
 import type { SelectionRangeEx } from 'roosterjs-editor-types';
 
 // In theory, all functions below are not necessary. We keep these functions here only for compatibility with old IEditor interface
@@ -84,31 +84,4 @@ export function convertDomSelectionToRangeEx(selection: DOMSelection | null): Se
                 areAllCollapsed: true,
             };
     }
-}
-
-/**
- * @internal
- * Create ranges from a table selection
- * @param selection The source table selection
- * @returns An array of DOM ranges of selected table cells
- */
-export function createTableRanges(selection: TableSelection): Range[] {
-    const result: Range[] = [];
-    const { table, firstColumn, firstRow, lastColumn, lastRow } = selection;
-    const cells = parseTableCells(table);
-
-    for (let row = firstRow; row <= lastRow; row++) {
-        for (let col = firstColumn; col <= lastColumn; col++) {
-            const td = cells[row]?.[col];
-
-            if (typeof td == 'object') {
-                const range = table.ownerDocument.createRange();
-
-                range.selectNode(td);
-                result.push(range);
-            }
-        }
-    }
-
-    return result;
 }
